timeit
